+++ sklearn/preprocessing/tests/test_target.py
@@ -4,6 +4,7 @@ from sklearn.utils.testing import assert_array_almost_equal
 from sklearn.utils.testing import assert_array_equal
 from sklearn.utils.testing import assert_equal
 from sklearn.utils.testing import assert_raises_regex
+from sklearn.utils.testing import assert_allclose
 
 from sklearn.preprocessing import TransformedTargetRegressor
 
@@ -19,59 +20,65 @@ def test_transformed_target_regressor_error_kwargs():
     X = friedman[0]
     y = friedman[1]
     # provide a transformer and functions at the same time
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     transformer=StandardScaler(),
-                                     func=np.exp, inverse_func=np.log)
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      transformer=StandardScaler(),
+                                      func=np.exp, inverse_func=np.log)
     assert_raises_regex(ValueError, "Both 'transformer' and functions"
                         " 'func'/'inverse_func' cannot be set at the"
-                        " same time.", clf.fit, X, y)
+                        " same time.", regr.fit, X, y)
 
 
 def test_transformed_target_regressor_invertible():
     X = friedman[0]
     y = friedman[1]
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     func=np.sqrt, inverse_func=np.log,
-                                     check_inverse=True)
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      func=np.sqrt, inverse_func=np.log,
+                                      check_inverse=True)
     assert_raises_regex(ValueError, "The provided functions or transformer"
                         " are not strictly inverse of each other.",
-                        clf.fit, X, y)
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     func=np.sqrt, inverse_func=np.log,
-                                     check_inverse=False)
+                        regr.fit, X, y)
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      func=np.sqrt, inverse_func=np.log,
+                                      check_inverse=False)
     # the transformer/functions are not checked to be invertible the fitting
     # should pass
-    clf.fit(X, y)
+    regr.fit(X, y)
 
 
 def test_transformed_target_regressor_friedman():
     X = friedman[0]
     y = friedman[1]
     # pass some functions
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     func=np.log, inverse_func=np.exp)
-    pred = clf.fit(X, y).predict(X)
-    y_tran = np.ravel(clf.transformer_.transform(y))
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      func=np.log, inverse_func=np.exp)
+    y_pred = regr.fit(X, y).predict(X)
+    y_tran = np.ravel(regr.transformer_.transform(y))
     assert_array_almost_equal(np.log(y), y_tran)
-    assert_array_almost_equal(y, np.ravel(clf.transformer_.inverse_transform(
+    assert_array_almost_equal(y, np.ravel(regr.transformer_.inverse_transform(
         y_tran.reshape(-1, 1))))
-    assert_equal(y.shape, pred.shape)
+    assert_equal(y.shape, y_pred.shape)
+    assert_allclose(y_pred, regr.inverse_func(regr.regressor_.predict(X)))
+    lr = LinearRegression().fit(X, regr.func(y))
+    assert_allclose(y_pred, regr.inverse_func(lr.predict(X)))
+    assert_array_equal(regr.regressor_.coef_.ravel(),
+                       lr.coef_.ravel())
     # pass a transformer
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     transformer=StandardScaler())
-    pred = clf.fit(X, y).predict(X)
-    assert_equal(y.shape, pred.shape)
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      transformer=StandardScaler())
+    y_pred = regr.fit(X, y).predict(X)
+    assert_equal(y.shape, y_pred.shape)
     y_mean = np.mean(y)
     y_std = np.std(y)
-    y_tran = np.ravel(clf.transformer_.transform(y.reshape(-1, 1)))
+    y_tran = np.ravel(regr.transformer_.transform(y.reshape(-1, 1)))
     assert_array_almost_equal((y - y_mean) / y_std, y_tran)
-    assert_array_almost_equal(y, np.ravel(clf.transformer_.inverse_transform(
+    assert_array_almost_equal(y, np.ravel(regr.transformer_.inverse_transform(
         y_tran.reshape(-1, 1))))
-    assert_equal(y.shape, pred.shape)
+    assert_equal(y.shape, y_pred.shape)
     lr = LinearRegression()
     ss = StandardScaler()
     lr.fit(X, ss.fit_transform(y[:, None])[:, 0])
-    assert_array_equal(clf.regressor_.coef_.ravel(),
+    assert_allclose(y_pred, ss.inverse_transform(lr.predict(X)))
+    assert_array_equal(regr.regressor_.coef_.ravel(),
                        lr.coef_.ravel())
 
 
@@ -80,22 +87,32 @@ def test_transformed_target_regressor_multioutput():
     y = friedman[1]
     y = np.vstack((y, y ** 2 + 1)).T
     # pass some functions
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     func=np.log, inverse_func=np.exp)
-    pred = clf.fit(X, y).predict(X)
-    y_tran = clf.transformer_.transform(y)
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      func=np.log, inverse_func=np.exp)
+    y_pred = regr.fit(X, y).predict(X)
+    y_tran = regr.transformer_.transform(y)
     assert_array_almost_equal(np.log(y), y_tran)
-    assert_array_almost_equal(y, clf.transformer_.inverse_transform(y_tran))
-    assert_equal(y.shape, pred.shape)
+    assert_array_almost_equal(y, regr.transformer_.inverse_transform(y_tran))
+    assert_equal(y.shape, y_pred.shape)
+    assert_allclose(y_pred, regr.inverse_func(regr.regressor_.predict(X)))
+    lr = LinearRegression().fit(X, regr.func(y))
+    assert_allclose(y_pred, regr.inverse_func(lr.predict(X)))
+    assert_array_equal(regr.regressor_.coef_.ravel(),
+                       lr.coef_.ravel())
     # pass a transformer
-    clf = TransformedTargetRegressor(regressor=LinearRegression(),
-                                     transformer=StandardScaler())
-    pred = clf.fit(X, y).predict(X)
-    assert_equal(y.shape, pred.shape)
+    regr = TransformedTargetRegressor(regressor=LinearRegression(),
+                                      transformer=StandardScaler())
+    y_pred = regr.fit(X, y).predict(X)
+    assert_equal(y.shape, y_pred.shape)
     y_mean = np.mean(y, axis=0)
     y_std = np.std(y, axis=0)
-    y_tran = clf.transformer_.transform(y)
+    y_tran = regr.transformer_.transform(y)
     assert_array_almost_equal((y - y_mean) / y_std, y_tran)
-    assert_array_almost_equal(y, clf.transformer_.inverse_transform(
+    assert_array_almost_equal(y, regr.transformer_.inverse_transform(
         y_tran))
-    assert_equal(y.shape, pred.shape)
+    assert_equal(y.shape, y_pred.shape)
+    ss = StandardScaler()
+    lr.fit(X, ss.fit_transform(y))
+    assert_allclose(y_pred, ss.inverse_transform(lr.predict(X)))
+    assert_array_equal(regr.regressor_.coef_.ravel(),
+                       lr.coef_.ravel())